+++ SandboxiePlus/SandMan/Windows/RecoveryWindow.cpp
@@ -481,7 +481,7 @@ void CRecoveryWindow::RecoverFiles(bool bBrowse, QString RecoveryFolder)
 	}
 
 
-	SB_PROGRESS Status = theGUI->RecoverFiles(FileList);
+	SB_PROGRESS Status = theGUI->RecoverFiles(m_pBox->GetName(), FileList);
 	if (Status.GetStatus() == OP_ASYNC)
 	{
 		connect(Status.GetValue().data(), SIGNAL(Finished()), this, SLOT(FindFiles()));
